OASIS/chat_application/backend/requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
python-socketio==5.8.0
cryptography==41.0.4
bcrypt==4.0.1
python-dotenv==1.0.0
PyJWT==2.10.1